OpenZFS: let userspace pointers be capabilities
Make variables that might point to userspace or be used as pointers in
userspace be capable of holding capabilities.  This means
 - changing uint64_t pointers to kuint64cap_t (in a world without hybrid
   kernels this would be the non-existant uint64ptr_t).
 - removing casts to uint64_t when assigning to such pointers to
   avoid losing provinance. Instead rely on implicit casts from
   uintptr_t to uint64_t, they should be fine even in 32-bit.
 - using uintcap_t casts in favor of uintptr_t casts when turning a
   kuint64cap_t into a pointer in the kernel to avoid losing capablity
   provinance in the CHERI case while still perfoming a cast through
   uintptr_t in the 32-bit case.

+++ sys/contrib/openzfs/lib/libzfs/libzfs_pool.c
@@ -1550,7 +1550,7 @@ zpool_destroy(zpool_handle_t *zhp, const char *log_str)
 		return (-1);
 
 	(void) strlcpy(zc.zc_name, zhp->zpool_name, sizeof (zc.zc_name));
-	zc.zc_history = (uint64_t)(uintptr_t)log_str;
+	zc.zc_history = (uintptr_t)log_str;
 
 	if (zfs_ioctl(hdl, ZFS_IOC_POOL_DESTROY, &zc) != 0) {
 		(void) snprintf(errbuf, sizeof (errbuf), dgettext(TEXT_DOMAIN,
@@ -1741,7 +1741,7 @@ zpool_export_common(zpool_handle_t *zhp, boolean_t force, boolean_t hardforce,
 	(void) strlcpy(zc.zc_name, zhp->zpool_name, sizeof (zc.zc_name));
 	zc.zc_cookie = force;
 	zc.zc_guid = hardforce;
-	zc.zc_history = (uint64_t)(uintptr_t)log_str;
+	zc.zc_history = (uintptr_t)log_str;
 
 	if (zfs_ioctl(zhp->zpool_hdl, ZFS_IOC_POOL_EXPORT, &zc) != 0) {
 		switch (errno) {
@@ -4398,7 +4398,7 @@ get_history(zpool_handle_t *zhp, char *buf, uint64_t *off, uint64_t *len)
 
 	(void) strlcpy(zc.zc_name, zhp->zpool_name, sizeof (zc.zc_name));
 
-	zc.zc_history = (uint64_t)(uintptr_t)buf;
+	zc.zc_history = (uintptr_t)buf;
 	zc.zc_history_len = *len;
 	zc.zc_history_offset = *off;
